% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_mve_sr_cpp}
\alias{compute_mve_sr_cpp}
\title{Compute Mean-Variance Efficient Portfolio Sharpe Ratio}
\usage{
compute_mve_sr_cpp(mu, sigma, selection, do_checks = FALSE)
}
\arguments{
\item{mu}{Mean vector.}

\item{sigma}{Covariance matrix.}

\item{selection}{Index vector with asset indices. Default is an empty
vector, which means all assets are selected.}

\item{do_checks}{Logical flag indicating whether to perform input checks (default = FALSE).}
}
\value{
A scalar value corresponding to \eqn{\sqrt{\mu^T \Sigma^{-1}\mu}}.
}
\description{
Given a mean vector \eqn{\mu}, a covariance matrix \eqn{\Sigma},
and an asset selection vector, this function computes the optimal Sharpe ratio defined as
\deqn{\sqrt{\mu^T \Sigma^{-1}\mu}}
over the selected assets.
If the provided asset selection vector has length less than N,
the function uses the subset of assets and computes the ratio using those assets.
}
