% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mve_weights.R
\name{compute_mve_weights}
\alias{compute_mve_weights}
\title{Compute MVE Weights}
\usage{
compute_mve_weights(mu, sigma, selection = c(), gamma = 1, do_checks = FALSE)
}
\arguments{
\item{mu}{First moment vector}

\item{sigma}{Covariance matrix}

\item{selection}{Asset selection vector (default = full selection)}

\item{gamma}{Risk aversion parameter (default = 1)}

\item{do_checks}{Logical flag to perform input checks (default = FALSE)}
}
\value{
A vector of portfolio weights.
}
\description{
Given the risk-aversion parameter \eqn{\gamma}, the first moment vector, the covariance matrix,
and the selection index vector, this function computes the mean variance efficient portfolio weights
\deqn{w = \frac{1}{\gamma}\, \text{sigma}^{-1}\, \mu},
over the selected assets.
If the provided asset selection vector has length less than N,
the function returns an N-length weight vector with zeros for the unselected assets.
}
\examples{
# Full selection example:
compute_mve_weights(mu = c(0.1, 0.2, 0.15),
            sigma = matrix(c(1, 0.2, 0.1,
                                     0.2, 1, 0.3,
                                     0.1, 0.3, 1), nrow = 3),
            selection = c(1, 2, 3),
            gamma = 1,
            do_checks = TRUE)

# Subset selection example:
compute_mve_weights(mu = c(0.1, 0.2, 0.15, 0.12),
            sigma = diag(4),
            selection = c(1, 3),
            gamma = 1,
            do_checks = TRUE)
}
