% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_mve_weights_cpp}
\alias{compute_mve_weights_cpp}
\title{Compute Mean-Variance Efficient (MVE) Portfolio Weights}
\usage{
compute_mve_weights_cpp(
  mu,
  second_moment,
  selection,
  gamma = 1,
  do_checks = FALSE
)
}
\arguments{
\item{mu}{First moment vector.}

\item{second_moment}{Second moment matrix.}

\item{selection}{Unsigned integer vector with asset indices.}

\item{gamma}{Risk aversion parameter. Default is 1.}

\item{do_checks}{Logical flag indicating whether to perform input checks (default = FALSE).}
}
\value{
An N-length vector of mean variance efficient weights.
}
\description{
Given a risk-aversion parameter \eqn{\gamma}, a first moment vector, a second moment matrix,
and a selection index vector, this function computes the mean variance efficient portfolio weights
\deqn{w = \frac{1}{\gamma}\, \text{second\_moment}^{-1}\, \text{first\_moment}},
over the selected assets.
If the provided asset selection vector has length less than N,
the function returns an N-length weight vector with zeros for the unselected assets.
}
