% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{.build_windows}
\alias{.build_windows}
\title{Build (in, out) index windows}
\usage{
.build_windows(
  Tobs,
  size_w_in,
  size_w_out,
  oos_type = c("rolling", "expanding")
)
}
\description{
Build (in, out) index windows
}
