% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factor_mom}
\alias{factor_mom}
\title{Momentum (tradable) factor.}
\format{
A matrix with \code{T} rows and 2 columns:
\describe{
\item{Date}{Date in yyyymm format.}
\item{Momentum}{Momentum factor (in decimals).}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
factor_mom
}
\description{
Monthly observations of the momentum (tradable) factor,
covering the period from 05/2008 to 12/2022.
}
\keyword{datasets}
