% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factors_ff5}
\alias{factors_ff5}
\title{Fama–French 5 (tradable) factors and risk-free rate.}
\format{
A matrix with \code{T} rows and 7 columns:
\describe{
\item{Date}{Date in yyyymm format.}
\item{...}{The five factors and the risk-free rate (in decimals).}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
factors_ff5
}
\description{
Monthly observations of the Fama–French 5 factors (tradable version),
covering the period from 05/2008 to 12/2022.
}
\keyword{datasets}
