% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{plot_empirical_suite_h1}
\alias{plot_empirical_suite_h1}
\title{Plot all standard empirics diagnostics from run_empirical_suite_h1()}
\usage{
plot_empirical_suite_h1(suite, fig_dir, stem_base)
}
\description{
Plot all standard empirics diagnostics from run_empirical_suite_h1()
}
