% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{plot_less_than_k}
\alias{plot_less_than_k}
\title{Plot fraction of windows with sparsity < k}
\usage{
plot_less_than_k(
  k_grid,
  less_than_k,
  labels = NULL,
  save_path = NULL,
  total_windows
)
}
\description{
Plot fraction of windows with sparsity < k
}
