% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{plot_selection_instability_empirics}
\alias{plot_selection_instability_empirics}
\title{Plot median selection instability vs k}
\usage{
plot_selection_instability_empirics(
  k_grid,
  sel_instab,
  labels = NULL,
  save_path = NULL
)
}
\description{
Plot median selection instability vs k
}
