% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{plot_turnover_empirics}
\alias{plot_turnover_empirics}
\title{Plot median turnover vs k}
\usage{
plot_turnover_empirics(
  k_grid,
  turnover,
  method_labels = NULL,
  save_path = NULL
)
}
\description{
Plot median turnover vs k
}
