% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_crsp}
\alias{returns_crsp}
\title{CRSP excess returns.}
\format{
A matrix with \code{T} rows and 201 columns:
\describe{
\item{Date}{Date in yyyymm format.}
\item{...}{Excess returns for each asset.}
}
}
\source{
\url{https://www.crsp.org}
}
\usage{
returns_crsp
}
\description{
Monthly CRSP excess returns for individual assets,
covering the period from 05/2008 to 12/2022.
}
\keyword{datasets}
