% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_ind17}
\alias{returns_ind17}
\title{Excess returns on 17 industry portfolios.}
\format{
A matrix with \code{T} rows and 18 columns:
\describe{
\item{Date}{Date in yyyymm format.}
\item{...}{Excess returns on the industry portfolios.}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
returns_ind17
}
\description{
Monthly excess returns for 17 industry portfolios,
covering the period from 05/2008 to 12/2022.
}
\keyword{datasets}
