% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_ind49_daily}
\alias{returns_ind49_daily}
\title{Daily excess returns: 49 industry portfolios (US)}
\format{
A data frame with columns:
\describe{
\item{DATE}{Integer date in YYYYMMDD format}
\item{Asset}{Excess returns}
}
}
\usage{
returns_ind49_daily
}
\description{
DATE plus excess returns (decimals) for 49 industry portfolios.
}
\keyword{datasets}
