% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_meac25}
\alias{returns_meac25}
\title{Excess returns on portfolios sorted on size and accruals.}
\format{
A matrix with \code{T} rows and 26 columns:
\describe{
\item{Date}{Date in yyyymm format.}
\item{...}{Excess returns on the portfolios.}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
returns_meac25
}
\description{
Monthly excess returns for 25 portfolios formed on size and accruals,
covering the period from 05/2008 to 12/2022.
}
\keyword{datasets}
