% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_meprior10_daily}
\alias{returns_meprior10_daily}
\title{Daily excess returns: ME×Prior 1–0 (US)}
\format{
A data frame with columns:
\describe{
\item{DATE}{Integer date in YYYYMMDD format}
\item{Asset}{Excess returns}
}
}
\usage{
returns_meprior10_daily
}
\description{
Daily excess returns: ME×Prior 1–0 (US)
}
\keyword{datasets}
