% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns_meprior122_daily}
\alias{returns_meprior122_daily}
\title{Daily excess returns: ME×Prior 12–2 (US)}
\format{
A data frame with columns:
\describe{
\item{DATE}{Integer date in YYYYMMDD format}
\item{Asset}{Excess returns}
}
}
\usage{
returns_meprior122_daily
}
\description{
Daily excess returns: ME×Prior 12–2 (US)
}
\keyword{datasets}
