% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf}
\alias{rf}
\title{Risk-free return.}
\format{
A matrix with \code{T} rows and 2 columns:
\describe{
\item{Date}{Date in yyyymm format.}
\item{RF}{Risk-free rate.}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
rf
}
\description{
Monthly risk-free return from the Fama–French datasets,
covering the period from 05/2008 to 12/2022.
}
\keyword{datasets}
