% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf_daily}
\alias{rf_daily}
\title{Daily risk-free rates (internal)}
\format{
A data frame with columns:
\describe{
\item{DATE}{Integer date in YYYYMMDD format}
\item{Asset}{RF}
}
}
\usage{
rf_daily
}
\description{
DATE and RF (decimals) aligned to the factors. Not exported.
}
\keyword{datasets}
