% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_mve_sr}
\alias{simulate_mve_sr}
\title{Simulate Sharpe Ratio Loss}
\usage{
simulate_mve_sr(mu, sigma, n_obs, max_card, max_comb = 0, do_checks = FALSE)
}
\arguments{
\item{mu}{A numeric vector; the population mean vector.}

\item{sigma}{A numeric matrix; the population covariance matrix.}

\item{n_obs}{An integer specifying the sample size to simulate.}

\item{max_card}{Maximum cardinality to consider (from 1 up to the number of assets).}

\item{max_comb}{Maximum number of combinations to consider. If 0 (default),
all combinations are computed.}

\item{do_checks}{Logical; if TRUE, input checks are performed.}
}
\value{
A list with:
\itemize{
\item \code{sample_mve_sr} computed as the optimal sample mve Sharpe ratio,
\item \code{sample_mve_sr_cardk} computed as the optimal sample mve Sharpe ratio
with cardinality \code{max_card},
\item \code{mve_sr_cardk_est_term} computed as \eqn{w^T \mu^T / \sqrt{w^T\sigma w}}
where \code{w} are the optimal sample mve weights,
\item \code{mve_sr_cardk_sel_term} computed as \eqn{\mu_S^T  \sigma_S^{-1} \mu_S}
where \code{S} is the set of assets yielding the optimal sample mve Sharpe ratio.
}
}
\description{
This function simulates a sample of size \code{n_obs} from a multivariate normal
distribution with mean vector \code{mu} and covariance matrix \code{sigma}.
It computes the sample mean vector (\code{mu_sample}) and the sample covariance matrix (\code{sigma_sample}),
then with the population and sample parameters,
the maximum cardinality (\code{max_card}), and the maximum number of combinations (\code{max_comb}),
it computes: the sample MVE Sharpe ratio, the sample MVE Sharpe ratio with
maximum cardinality k, and the population MVE Sharpe ratio with maximum
cardinality k decomposed in estimation and selection component.
}
\examples{
\dontrun{
# Example with three assets:
mu <- c(0.1, 0.2, 0.15)
  sigma <- diag(3)
  mve_sr <- 0.8
  n_obs <- 100
  max_card <- 2
  result <- simulate_mve_sr(mu,
                            sigma,
                            n_obs,
                            max_card,
                            do_checks = TRUE)
  print(result)
}
}
