% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{sps_make_k_grid}
\alias{sps_make_k_grid}
\title{Build a k-grid with basic checks}
\usage{
sps_make_k_grid(N, k_min = 1L, k_step = 3L, k_cap = NULL)
}
\description{
Build a k-grid with basic checks
}
