% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{sps_make_output_dirs}
\alias{sps_make_output_dirs}
\title{Create output directories for an empirics run}
\usage{
sps_make_output_dirs(out_dir, fig_dir)
}
\description{
Create output directories for an empirics run
}
