% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{sps_make_stem_base}
\alias{sps_make_stem_base}
\title{Standard filename stem for an empirics run}
\usage{
sps_make_stem_base(
  method,
  refit = FALSE,
  panel_tag,
  factors_tag,
  mkt_tag,
  N,
  h = 1L,
  normalize_weights = NULL
)
}
\description{
Standard filename stem for an empirics run
}
