% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{sps_stack_metric}
\alias{sps_stack_metric}
\title{Stack a metric from a list of summary data.frames into a K x M matrix}
\usage{
sps_stack_metric(summaries, metric_col, k_grid)
}
\description{
Stack a metric from a list of summary data.frames into a K x M matrix
}
