% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_empirics.R
\name{sps_subset_panel}
\alias{sps_subset_panel}
\title{Subset a panel to the first N assets (and align rf)}
\usage{
sps_subset_panel(R_excess, rf = NULL, n_assets = NULL)
}
\description{
Subset a panel to the first N assets (and align rf)
}
