% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mve_sr.R
\name{compute_mve_sr}
\alias{compute_mve_sr}
\title{Compute Optimal Portfolio Sharpe Ratio}
\usage{
compute_mve_sr(mu, sigma, selection = c(), do_checks = FALSE)
}
\arguments{
\item{mu}{Mean vector.}

\item{sigma}{Covariance matrix.}

\item{selection}{Asset selection vector (default = full selection).}

\item{do_checks}{Logical flag to perform input checks (default = FALSE).}
}
\value{
A scalar value corresponding to \eqn{\sqrt{\mu^T \Sigma^{-1}\mu}}.
}
\description{
Given a mean vector \eqn{\mu}, a covariance matrix \eqn{\Sigma},
and an asset selection vector, this function computes the optimal square
Sharpe ratio defined as
\deqn{\sqrt{\mu^T \Sigma^{-1}\mu}}
over the selected assets.
If the provided asset selection vector has length less than N,
the function uses the subset of assets and computes the ratio using those assets.
}
\examples{
# Full selection example:
compute_mve_sr(mu = c(0.1, 0.2, 0.15),
                       sigma = matrix(c(1, 0.2, 0.1,
                                          0.2, 1, 0.3,
                                          0.1, 0.3, 1), nrow = 3),
                       selection = c(1, 2, 3),
                       do_checks = TRUE)

# Subset selection example:
compute_mve_sr(mu = c(0.1, 0.2, 0.15, 0.12),
                       sigma = diag(4),
                       selection = c(1, 3),
                       do_checks = TRUE)
}
