% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_sr_cpp}
\alias{compute_sr_cpp}
\title{Compute Portfolio Sharpe Ratio}
\usage{
compute_sr_cpp(weights, mu, sigma, selection, do_checks = FALSE)
}
\arguments{
\item{weights}{A numeric vector of portfolio weights.}

\item{mu}{A numeric vector of expected returns.}

\item{sigma}{A numeric covariance matrix.}

\item{selection}{Index vector with asset indices. Default is an empty
vector, which means all assets are selected.}

\item{do_checks}{Logical flag indicating whether to perform input checks (default = false).}
}
\value{
A double representing the Sharpe ratio.
}
\description{
Given the mean vector \eqn{\mu}, the covariance matrix \eqn{\Sigma},
and the asset selection vector, this function computes the Sharpe ratio defined as
\deqn{\frac{w^T \mu}{\sqrt{w^T \Sigma w}}}
over the selected assets.
}
